Avoid overflow errors on statistics describing large arrays.

+++ cuda/intro-vtx-transform/vtx-xform-size.cu
@@ -678,10 +678,10 @@ main(int argc, char **argv)
       exit(1);
     }
 
-  const int in_size_elts = app.num_vecs * N;
-  const int in_size_bytes = in_size_elts * sizeof( app.h_in[0] );
-  const int out_size_elts = app.num_vecs * M;
-  const int out_size_bytes = out_size_elts * sizeof( app.h_out[0] );
+  const size_t in_size_elts = size_t(app.num_vecs) * N;
+  const size_t in_size_bytes = in_size_elts * sizeof( app.h_in[0] );
+  const size_t out_size_elts = size_t(app.num_vecs) * M;
+  const size_t out_size_bytes = out_size_elts * sizeof( app.h_out[0] );
 
   const int overrun_size_elts = thd_per_block_goal * max(N,M);
   const int overrun_size_bytes = overrun_size_elts * sizeof( app.h_out[0] );
@@ -728,7 +728,7 @@ main(int argc, char **argv)
   const int64_t num_ops = int64_t(M) * N * app.num_vecs;  // Multiply-adds.
 
   // Amount of data in and out of GPU chip.
-  const int amt_data_bytes = in_size_bytes + out_size_bytes;
+  const int64_t amt_data_bytes = in_size_bytes + out_size_bytes;
 
   double elapsed_time_s = 86400; // Reassigned to minimum run time.
